refactor: move add config version function
